+++ routes/web.php
@@ -200,26 +200,42 @@ Route::prefix('admin')->middleware(['auth', 'prevent.guest.admin'])->name('admin
 
     Route::prefix('finance')->name('finance.')->group(function () {
         Route::prefix('years')->name('years.')->group(function () {
-            Route::get('/', [FinanceYearController::class, 'list'])->name('list');
-            Route::post('/', [FinanceYearController::class, 'store'])->name('store');
-            Route::get('/get-by-year', [FinanceYearController::class, 'getYearByNumber'])->name('getByYear');
-            Route::get('/{id}', [FinanceYearController::class, 'show'])->name('show')->where('id', '[0-9]+');
-            Route::put('/{id}/target', [FinanceYearController::class, 'updateTarget'])->name('updateTarget')->where('id', '[0-9]+');
-            Route::put('/{id}/note', [FinanceYearController::class, 'updateNote'])->name('updateNote')->where('id', '[0-9]+');
-            
+            // Read permissions
+            Route::get('/', [FinanceYearController::class, 'list'])->name('list')->middleware('permission:finance.read');
+            Route::get('/get-by-year', [FinanceYearController::class, 'getYearByNumber'])->name('getByYear')->middleware('permission:finance.read');
+            Route::get('/{id}', [FinanceYearController::class, 'show'])->name('show')->where('id', '[0-9]+')->middleware('permission:finance.read');
             Route::get('/{yearId}/months/{month}', [FinanceMonthController::class, 'show'])
                 ->name('months.show')
-                ->where(['yearId' => '[0-9]+', 'month' => '[1-9]|1[0-2]']);
+                ->where(['yearId' => '[0-9]+', 'month' => '[1-9]|1[0-2]'])
+                ->middleware('permission:finance.read');
+
+            // Create permissions
+            Route::post('/', [FinanceYearController::class, 'store'])->name('store')->middleware('permission:finance.create');
+
+            // Update permissions
+            Route::put('/{id}/target', [FinanceYearController::class, 'updateTarget'])->name('updateTarget')->where('id', '[0-9]+')->middleware('permission:finance.update');
+            Route::put('/{id}/note', [FinanceYearController::class, 'updateNote'])->name('updateNote')->where('id', '[0-9]+')->middleware('permission:finance.update');
         });
         
         Route::prefix('months')->name('months.')->group(function () {
-            Route::put('/{monthId}', [FinanceMonthController::class, 'update'])->name('update')->where('monthId', '[0-9]+');
+            // Read permissions
+            Route::get('/{monthId}/day-details', [FinanceMonthController::class, 'dayDetails'])->name('dayDetails')->where('monthId', '[0-9]+')->middleware('permission:finance.read');
             Route::prefix('{monthId}/expenses')->name('expenses.')->group(function () {
-                Route::get('/', [FinanceMonthController::class, 'getExpenses'])->name('index')->where('monthId', '[0-9]+');
-                Route::post('/', [FinanceMonthController::class, 'storeExpense'])->name('store')->where('monthId', '[0-9]+');
-                Route::put('/{id}', [FinanceMonthController::class, 'updateExpense'])->name('update')->where(['monthId' => '[0-9]+', 'id' => '[0-9]+']);
-                Route::delete('/{id}', [FinanceMonthController::class, 'destroyExpense'])->name('destroy')->where(['monthId' => '[0-9]+', 'id' => '[0-9]+']);
+                Route::get('/', [FinanceMonthController::class, 'getExpenses'])->name('index')->where('monthId', '[0-9]+')->middleware('permission:finance.read');
+                
+                // Create permissions
+                Route::post('/', [FinanceMonthController::class, 'storeExpense'])->name('store')->where('monthId', '[0-9]+')->middleware('permission:finance.create');
+                
+                // Update permissions
+                Route::put('/{id}', [FinanceMonthController::class, 'updateExpense'])->name('update')->where(['monthId' => '[0-9]+', 'id' => '[0-9]+'])->middleware('permission:finance.update');
+                
+                // Delete permissions
+                Route::delete('/{id}', [FinanceMonthController::class, 'destroyExpense'])->name('destroy')->where(['monthId' => '[0-9]+', 'id' => '[0-9]+'])->middleware('permission:finance.delete');
             });
+
+            // Update permissions
+            Route::put('/{monthId}', [FinanceMonthController::class, 'update'])->name('update')->where('monthId', '[0-9]+')->middleware('permission:finance.update');
+            Route::post('/{monthId}/lock', [FinanceMonthController::class, 'lock'])->name('lock')->where('monthId', '[0-9]+')->middleware('permission:finance.update');
         });
 
     });
